fix bb 3

+++ src/components/leaf-effect.tsx
@@ -9,7 +9,7 @@ import { inSphere } from 'maath/random';
 function FallingLeaves() {
     const ref = useRef<THREE.Points | null>(null);
     const count = 1000;
-    const [positions] = useState(() => inSphere(new Float32Array(count * 3), { radius: 5 }));
+    const [positions] = useState(() => inSphere(new Float32Array(count * 3), { radius: 5 }) as Float32Array);
 
     useFrame((state, delta) => {
         if (ref.current) {
